Fix issue 1B2K#31 move unused screens to folder
