+++ src/screens/not-using/CalendarScreen.js
@@ -8,15 +8,12 @@ import {
   SafeAreaView,
 } from "react-native";
 
-import { AuthContext } from "../config/context";
-import colors from "../config/colors";
+import colors from "../../config/colors";
 
-const SignUpScreen = () => {
-  const { signUp } = React.useContext(AuthContext);
+const CalendarScreen = () => {
   return (
     <SafeAreaView style={styles.container}>
-      <Text style={styles.text}>Hello This is the Sign Up area</Text>
-      <Button title="Sign Up" onPress={() => signUp()} />
+      <Text style={styles.text}> Hello this is calendar</Text>
     </SafeAreaView>
   );
 };
@@ -30,4 +27,5 @@ const styles = StyleSheet.create({
     textAlign: "center",
   },
 });
-export default SignUpScreen;
+
+export default CalendarScreen;
